api/cache: fix panic
Fixes:

    panic: runtime error: invalid memory address or nil pointer dereference
    [signal SIGSEGV: segmentation violation code=0x1 addr=0x10 pc=0x1b5446c]

    goroutine 1 [running]:
    koding/klientctl/endpoint/auth.(*Client).Delete(0x27ab9a0, 0x0,
    0x27557a0, 0xc42021fcc0)
        /Users/rjeczalik/src/github.com/koding/koding/go/src/koding/klientctl/endpoint/auth/auth.go:160+0x7c

+++ go/src/koding/api/cache.go
@@ -97,7 +97,7 @@ func (s *SessionCache) Auth(opts *AuthOptions) (*Session, error) {
 	// fail are filesystem errors - there is no recovery from
 	// that, and falling back to memory makes sense only for
 	// Klient.
-	if opts.Refresh {
+	if opts.Refresh && session != nil {
 		_ = s.Storage.Delete(session)
 	}
 	if err == nil {
